Merge 117146 - Remove contexts from lists of shared contexts on lost context.
Even if a context was lost, we tried to use it as the share group for newly-created contexts, resulting in a new context that shared with nothing. Remove all the dead contexts from the list of shared contexts to avoid that.


BUG=107374,107795
TEST=


Review URL: http://codereview.chromium.org/9169026

TBR=jbauman@chromium.org
Review URL: http://codereview.chromium.org/9185035