Fix flakiness and re-enable http search in find_in_page.FindMatchTests.testSearchInPDF for 963
Add loop to _VerifySearchInPDFURL to try FindInPage 10 times and re-raise
the exception on the final loop. This is required because
_GetResultFromJSONRequest appears to timeout if called before the PDF has
loaded.

Add timeout parameter to pyauto.FindInPage to be passed to
_GetResultFromJSONRequest to speed up the loop in _VerifySearchInPDFURL
from the default of 45 seconds to 1 second.

Re-enable test on Win, Mac, and Linux.

Patch by Victoria Clarke <vclarke@chromium.org>

BUG=107448, 70928, 93956, 91225
Review URL: http://codereview.chromium.org/8985003