Implement blocking read. Reading was working fine for synchronous case, but the blocking read callback was being turned into an event that was simply dropped onto the floor in schema_generated_bindings.js. And the functional test wasn't even setting an onEvent handler had the event not been dropped on the floor.
To reduce the likelihood of the unhandled event happening again, I've strengthened the event-handler wrapper in schema_generated_bindings.js to spit out a log message if it encounters an event type that it doesn't recognize. I've also written correct functional/unit tests for the specific case that was previously ignored.

Finally, the test echo server now handles an arbitrary number of echo round-trips, and the protocol it uses for line endings and when to quit is expressed in terms of simple string patterns.

BUG=109337
TEST=improved


Review URL: http://codereview.chromium.org/9186014