Keep NaCl toolchain tarballs, so we can do a faster clean.
Add the --keep argument to the downloader to that we keep the original
tarball instead of deleting it after untarring.  Currently we only blow
away the toolchain when we download a new one.  Since it lives in the
sources, and since we add components to it during the build process,
the toolchain does not get cleaned through a conventional clean such as
deleting out/Debug.

The SDK builders will build an SDK tarball which contains the toolchain
plus user libraries.  To improve buildtime we will want to re-untar
instead of downloading everytime.

BUG=109207
Review URL: http://codereview.chromium.org/9186054