Content: First draft at a content_shell on GTK.
THIS IS NOT READY FOR DAILY USE.

The shell, in its current state, can bring up the default google.com
homepage. And that's about it. There are problems with the url bar,
scrolling (weird crashes complaining about not having a
ResourceBundle), it breaks badly on popups, et cetera.

BUG=90445
TEST=none; this is a testing tool.


Review URL: http://codereview.chromium.org/9213012