Captured cookies WEBUI improved button spacing and layout. Sets dialog title.
Rather than bottom-justifing the buttons so that there's no ugly space after
them when the info-area isn't expanded, I've made it so that the two tree
components are slightly expanded when the info-area isn't expanded so that
the bottons are still justified along the bottom.  This approach doesn't 
waste any screen space.  All buttons are left-justified, except for the close
button which appears on the right.  These changes work both RTL and LTR.

This is a replacement for CL 8968029.

BUG=104494
TEST=manually

Review URL: http://codereview.chromium.org/9207010