[Sync] Pave the way for fewer asynchronous calls and UI spin while setting a passphrase
As of today, when a user sets a passphrase, SetPassphrase() is invoked
from the UI thread, but the actual work is performed on the syncer
thread. This means that when the UI thread sets a decryption passphrase,
it doesn't actually know right away whether the passphrase was accepted
or not, resulting in the user having to see a spinner during this time.

This patch does the following:

- Changes the OnPassphraseRequired API to include a
  sync_pb::EncryptedData parameter, that contains a copy of the
  cryptographer's pending keys when the reason is REASON_DECRYPTION.
- Adds the cached_pending_keys_ data member to ProfileSyncService, so
  that it can cache the cryptographer's pending keys when
  OnPassphraseRequired is called.
- Adds a new API called CheckPassphraseAgainstCachedPendingKeys to
  ProfileSyncService that can be called by the UI to first try to decrypt the
  cached pending keys with the passphrase that was entered, and if this
  fails, directly bubble up an incorrect passphrase dialog instead of
  showing the user a spinner.

BUG=108718
TEST=try to enter an incorrect passphrase; try to enter an old
     passphrase, sync_integration_tests

Review URL: http://codereview.chromium.org/9107027