Mac/Aura compositor_unittests are failing
The problem was that the ui::Compositor keeps a "delegate" reference to the
TestCompositorHostMac, but that gets deleted prior to the ref-counted
ui::Compositor.  The ui::Compositor reference is being held by the
AcceleratedTestView which de-ref's the compositor too late in the game, namely
when the Mac autorelease pools get drained.

This is resolved here by explicitly clearing the reference that the
view holds to the compositor.

BUG=110666
TEST=LayerWithRealCompositorTest.*
R=piman@chromium.org

Review URL: https://chromiumcodereview.appspot.com/9253025