Change grow box computation back to a method on BrowserWindow
This is a revert of http://codereview.chromium.org/3547008, which was
added to implement sidebars. aa has reverted most of the sidebar code
( http://codereview.chromium.org/9006027/ ), so this is no longer needed.

This not only makes the code simpler (look at all the red), but also fixes a
long-standing regression on Mac that was caused by the CL this reverts.

Parts of the windows part with help form sky@.

This does not add back the code in rwhv_win that was added by mad@ for
bug 458, since it has since then bitrotted, and it's not used anyway. If someone
wants to attack 458 again,
http://codereview.chromium.org/16488/diff/1101/chrome/browser/renderer_host/render_widget_host_view_win.cc
should be fairly easy to add back.

BUG=458,70482,107646
TEST=Not too much stuff breaks; resize box looks good again on 10.5/10.6
TBR=avi

Review URL: http://codereview.chromium.org/9265018