Aura: Fix crash on omnibox input in compact window mode
The layout manager for compact window mode calls Show() on the status area widget as part of its visibility updates when a new Aura window opens.  This causes the status area widget to try to activate and take focus, which tries to release focus and close the omnibox popup, which causes a crash.  Only call Show() if the widget is currently invisible, for transitions out of fullscreen.

BUG=110897
TEST=Run Aura in compact window mode, type into omnibox, should not crash

Review URL: https://chromiumcodereview.appspot.com/9160002