Copy NMF unmodified from 'site-scons' to tools
Add a script for generating NMF files to the tools directory of the SDK.
This already exists and is copied from another location in the SDK as-is.
This is done as a copy instead of a move so that the scons components in
the sdk can be removed as one atomic commit.

BUG= 111224
R= bradnelson@chromium.org
Review URL: http://codereview.chromium.org/9284032