Use type 'none' instead of 'settings', as the latter is invalid in gyp.
gyp currently warns about this, but I intend to make it an error.
Due to the warning this had already been fixed in most of the Chrome tree,
but not in some of the use_system_foo blocks.

(DEPS change: roll ICU forward to pick up same fix.)


Review URL: http://codereview.chromium.org/9287001