Enhance the PPAPI enter tracking.
The goal here is to be able to add thread checking to the enter functions, which will require more types of failure than just "bad resource".

I folded completion callback force executing into this, which also simplifies the thunks a bit. If we like this method, I'll convert the rest and delete MayForceCallback.

I converted URLLoader and Graphics2D to use this.

Review URL: https://chromiumcodereview.appspot.com/9235003