Reduce chrome/ dependencies on platform-specific RenderWidgetHostView implementations.
Various dependencies on the GTK version remain, as does a single,
probably easy-to-remove dependency on the Mac implementation.

Did this primarily as follows:

a) Enforce the intent, that RenderWidgetHostView instances are created
via factory function.

b) Migrated a couple of methods to the interface, skipping
implementation for platforms where not needed.

c) Added InitAsChild, with parent view parameter, to the interface so
that Windows initialization can go through interface.  Incidentally,
this fixes bug 102450.  As noted in the bug, the approach of passing a
parent view pointer is somewhat ugly as it isn't needed by all
platform-specific implementations, but I think it is acceptable.

TBR=owners (interface change LGTMed by owner, rest is trivial updates)

BUG=98716,102450

Review URL: https://chromiumcodereview.appspot.com/9281003