linux: Make PlatformFontPango handle family lists.
A PangoFontDescription can hold a comma-separated list of
font family names (and we currently use lists for Chrome
OS), but PlatformFontPango was just passing the whole string
to FontConfig.  This change makes us first split the family
on commas and then pass the resulting values to FontConfig.

BUG=107389,109961
TEST=manual: tested that we use Chrome Droid Sans on Chrome OS instead of falling back to sans all the time


Review URL: http://codereview.chromium.org/9288009