Fixes for the net-internals StopCapturing button.
It now dumps a log to memory and loads it.  The
results in updating all tabs, hiding tabs and buttons
that don't work when the BrowserBridge is disabled.

Also allows export of logs in stop capturing mode or
when viewing a loaded log (Which are now equivalent
states).

R=eroman@chromium.org
BUG=111195

Review URL: http://codereview.chromium.org/9291001