Reland 119457: --test-gl-lib for GpuPixelBrowserTests
This CL causes the GpuPixelBrowserTests to load libllvmpipe.so (if it is available) and fallback to standard libGL.so otherwise. This is so that we can do a smoke test of the GPU accelerated display path on the Aura dbg bot on the Chromium waterfall.

BUG=none
TEST=none

Review URL: https://chromiumcodereview.appspot.com/9298022