Defer render_widget draw until host window is available
In accelerated compositing mode we can't start rendering until the
native window (HWND/gtk widget) are created on Windows and GTK without
aura. For most render_widgets the code today is racy, but for
window.open() and similar creation paths this is a race that we pretty
much always lose. This defers the first DoDeferredUpdate on a render_widget
until its host_window_ is available.

Also defers querying for GL_CHROMIUM_supports_swapbuffers_callback
extension until the first draw, since the context isn't ready until the
host window is available.

BUG=106815
TEST=start with --force-compositing-mode and --show-fps-counter, load up
a page that uses window.open() to open a popup, and confirm that the
popup is in compositing mode


Review URL: http://codereview.chromium.org/9225050