Quick-and-dirty hacks for getting policy working behind proxies.
This is check-in for mattias.
Original review: http://codereview.chromium.org/9315011/

Note that this hard-wires policy OAuth fetching in login_utils.cc to the
authentication profile. I'm not entirely sure this is a good idea, but
required since the user profile isn't in a state to allow successful
requests at this point (the use-shared-proxies flag is off). Other stuff
in login_utils.cc might need the hack too (sync?).

BUG=chromium-os:25703
TEST=Verify fix for chromium-os:25703
TBR=mnissler@chromium.org,zelidrag@chromium.org,sumit@chromium.org,pastarmovj@chromium.org,kuan@chromium.org

Review URL: https://chromiumcodereview.appspot.com/9309024