Make chrome://omnibox/ be able to report per-provider matches.
This will be very useful for those of us working
on improving the scoring/ranking of autocomplete
suggestions.

The presentation method could probably be improved, but I'd prefer
to check this in as-is, and see how us omnibox developers use it in
practice.  Then we'll have a better idea of how chrome://omnibox/
should convey the information we most need.

BUG=111791
TEST=


Review URL: http://codereview.chromium.org/9289028