Revert 120150 (due to compile failure on chromiumos aura) - [Sync] Add ChromeSyncNotificationBridge class to sync notifier framework for sessions use.
The ChromeSyncNotificationBridge listens to notifications from chrome events that should
trigger a sync notification. It is a thread-safe class that will post the
notifications to the thread the SyncNotifierObserver is on, thereby allowing,
for example, events on the UI thread to trigger notifications on the Sync
thread. At the moment the only such event is NOTIFICATION_SYNC_REFRESH, which
is only triggered by the sessions datatype.

BUG=103469
TEST=manually going to chrome://newtab#opentabs and ensuring a sync cycle
occurs.

Review URL: http://codereview.chromium.org/9079002

TBR=zea@chromium.org
Review URL: https://chromiumcodereview.appspot.com/9316054