[Sync] Fix DCHECK() following AwaitQuiescence() failure
Basically WaitUntilTimestampMatches() should be the one that sets
timestamp_match_partner_ to NULL, as AwaitStatusChangeWithTimeout()
only does so when the status change happens (i.e., not when it times
out).

Re-enable migration tests that were running into this DCHECK. (The
original problem might still be there, but I was unable to repro
locally, so it may be fixed.  If it still fails, I will investigate
further.)

BUG=100382
TEST=


Review URL: http://codereview.chromium.org/9309071