Re-enable newly de-flaked test.
This test was disabled because it became flaky and eventually continually timed
out. After investigation, including being able to repro the original problem in
an XP VM, I tracked the problem back to a race condition where the ESC that
closes the dialog can sometimes cause that dialog to be closed before the ESC
event ACK is sent back, thus the test was hanging until timeout. That is filed
as http://code.google.com/p/chromium/issues/detail?id=111269 but it's low
priority since it's really a test-only issue (and only for this test AFAICT).

This test will still fail on Windows if the browser isn't the front window
(there are a number of browser and interactive tests with this issue; see
http://code.google.com/p/chromium/issues/detail?id=106498). For this test part
of it is that the SendKeyPress on Windows doesn't target a window with the
event - the event always goes to the front window, even if that's not the
correct place.

BUG=94864, 111269
TEST=Crank up a slow XP VM, load up the CPU (I used HeavyLoad), run the test
lots (I used a repeat of 100). Before clobber, problem almost right
away. After, no problem.

Review URL: http://codereview.chromium.org/9310016