Revert 120387 - Don't try to collect memory details during shutdown.
There seems to be cases where the asynchronous task of collecting memory details is executed during shutdown.

This causes an attempt at using the profile object set in the RenderProcessHost after the profile was destroyed.

To avoid the crash, we simply interrupt the memory details collection if we identify that we are trying to shutdown.

BUG=112383
TEST=Make sure we still properly collect memory details.

Review URL: http://codereview.chromium.org/9317058

TBR=mad@chromium.org
Review URL: https://chromiumcodereview.appspot.com/9317090