Delete host.json and exit if Host ID deleted from server
Follows on from http://codereview.chromium.org/9303008/.  This looks at the
exit-code from remoting_me2me_host.  If the reason for termination was that the
Host ID is no longer valid on the server, this deletes the host.json and
terminates the running daemon (shutting down the X session).

BUG=110046
TEST=Manual

Review URL: https://chromiumcodereview.appspot.com/9322001