Fix crash on disconnect
This change wraps the end of the input pipeline (ClientEventDispatcher) in a
forwarder (newly introduced InputFilter). This abstraction relieves
ChromotingInstance, and everything else, from the awkward and (provably)
bug-prone task of tracking the component's lifetime. It also allows the input
pipeline to be initialized alongside the connection, making for a simpler
contract and providing better code readability, among other things.
    
BUG=112116


Review URL: http://codereview.chromium.org/9320047