Introduce ResourceThrottle as a replacement for chrome's usage of ResourceHandler.
ResourceThrottle provides a subset of the functionality of ResourceHandler.  It can be used to defer the start, redirect and read phases of loading a resource.  A ResourceThrottle is initialized with a ResourceThrottle::Controller that provides an interface for the throttle implementation to resume or cancel a deferred load.

ResourceThrottle instances have their lifetime bound to the URLRequest they are associated with.  This eliminates the need for a method like ResourceHandler::OnRequestClosed.  To support this change, ResourceThrottle implementations utilize WeakPtr (e.g., when querying the SafeBrowsingService, a WeakPtr back to the SafeBrowsingResourceThrottle is bound to the UrlCheckCallback).  This simplifies the ResourceThrottle implementations as they don't have to worry about receiving callbacks after the URLRequest has been deleted.
Review URL: https://chromiumcodereview.appspot.com/9008080