UI changes for better session restore.
When the "reopen last pages" setting is selected, disable
the "clear on exit" setting and the session only cookie
setting.

If the user has the "reopen last pages" setting, display a
dialog when the user opens the settings page.

Change the label of the "reopen last pages" setting.

These changes are behind the command line flag
--enable-restore-session-state.

BUG=104289
TEST=NONE


Review URL: http://codereview.chromium.org/9235058