Fix incomplete teardown of demuxer.
I ran into an issue with PipelineIntegrationTests where one of the
security videos was causing ~FileDataSource to DCHECK complain about
file_ never being torn down.

Tracing this back, it's because we return NULL for the demuxer when
an error is encountered. This prevents demuxer->Stop() from being called
which leaves the data source in an unhappy state.

BUG=none
TEST=PipelineIntegrationTests/media_unittests


Review URL: http://codereview.chromium.org/9350004