Provide a hint for the idle notification handler.
The IdleNotification() function accepts an optional hint argument since V8 3.8.1 (rolled in r114984). 

This argument roughly specifies the expected duration of the idle pause and thus defines the amount of work to be done in the function. There is no guarantee that the actual work will match the hint though.

This CL provides hints for idle notifications sent from the foreground tab depending on the idle timer delay. This somewhat reduces idle notification pauses at the beginning of "idleness" period.

Review URL: http://codereview.chromium.org/9244007