Adds support for the credentials frame.
Need to allow spdy sessions to see if the underlying socket is required
OBCs. If so, and if pooling a connection, then we need to generate a
CREDENTIALS frame for this origin.

Need to add support for the RST_FRAME status code of INVALID CREDENTIALS

Review URL: http://codereview.chromium.org/9294001