Add a switch to invert web contents (non-accel only).
The switch will just be used during development of this feature.
The idea is for this to be turned automatically for Windows users
who enable an inverted (white-on-black) color scheme - see bug
for detail.

This first patch only handles the non-accelerated graphics path;
a subsequent change will handle the accelerated path.

BUG=112944
TEST=manually test with --invert-web-contents switch


Review URL: http://codereview.chromium.org/9353012