[Sync] Make Bookmark datatype more resilient against bookmark model failures.
We now trigger an unrecoverable error either during association or when
applying changes if the bookmark model is unable to create a new bookmark node.
The error will contain the bookmark title and url we were trying, and display
the information in about:sync, hopefully giving us more insight into when
this kind of thing occurs (and avoiding the crashes that happen currently).

BUG=108978
TEST=


Review URL: http://codereview.chromium.org/9316108