[uber] forward mouse wheels on uber-frame to showing subpage
although this is imperfect because the scroll amount may be slightly wrong, it is the least of all evils. I even tried to refactor the iframes so that the nav bar was in every iframe, and no event forwarding would be necessary, but this created new hells: inter-page transition animations and code duplication and complexification that caused me to beat a hasty retreat.

BUG=none
TEST=using scroll wheel while mouse is over nav area works. More or less.

Review URL: https://chromiumcodereview.appspot.com/9348010