ash: Fix creating the modal screen.
When the modal screen is animating on its way to hide itself, if another call to
CreateModalScreen is called, then it doesn't stop the hiding process. So make
sure that when creating the modal-screen, and it already exists, it is
explicitly told to Show.

BUG=113139
TEST=see bug


Review URL: http://codereview.chromium.org/9365018