Create a 4-bucket partial pre-read experiment.
This CL creates groups for a series of 4-bucket partial pre-read experiments.

On Stable, which has a large reporting user population a given build might report into the following buckets:

- 97% of users get the curent default behavior (100% pre-read)
- 1% of users will pre-read 25%
- 1% of users will pre-read 40%
- 1% of users will pre-read 0%

The 0 and 100 buckets represent our controls, and the 25 and 40 are the experimental values. The experimental values would be varied over successive builds to explore the pre-read-percentage space.

For channels with smaller reporting populations, the bucket sizes will be 25% of the population each.

Note that The UMA group labels for the above replace the current PreReadEnabled/Disabled labels.

BUG=chromium:98508
TEST=

Review URL: http://codereview.chromium.org/9352001