Enable devtools via context menu in platform apps and popup extensions.
This is useful to allow developers of apps and extensions to easily use
the devtools in platform apps and popup extensions.

To enable devtools to work on popup extensions a new notification
NOTIFICATION_DEVTOOLS_WINDOW_OPENING has been added. This lets the popup
become aware it is being debugged, which allows it to stay up when it 
loses focus (e.g. when the devtools are clicked).

BUG=53518,111017
TEST=Manual testing on Windows that dev tools are usable via the context
menu in these situations, and they are still disabled when they should
be.


Review URL: http://codereview.chromium.org/9235002