Ensure that we record sync promo suppressions
Currently we record which version of the sync promo a user was shown from the SyncPromoUI constructor.

This lead to a problem where we didn't count the "don't show sync promo" trial. To fix this we explicitly record the "don't show" trial in BrowserInit.

BUG=113135
TEST=


Review URL: http://codereview.chromium.org/9358017