Revert 121165 - Revert 121068 - linux: add a use_gold flag that defaults to on
[This CL was backed out speculatively due to failures on the cros trunk bot. But this CL
was not the cause of that problem, so I'm un-doing the revert.]

Currently, gyp attempts to autodetect whether the linker supports
the --icf flag.  It's more predictable to just require the caller
to specify whether they're using gold.

As a first demonstration of this, drop LINKER_USES_ICF code in the
Chrome build.  (I'll follow up with a gyp patch to remove the
LINKER_USES_ICF stuff in gyp.)

Review URL: https://chromiumcodereview.appspot.com/9368015

TBR=evan@chromium.org
Review URL: https://chromiumcodereview.appspot.com/9372012