Simplify logic in SyncScheduler::ScheduleNextSync
The else branch, which was intended to handle the first in a series of
failures, was almost impossible to hit.  The reason for this is that the
'else if' clause that preceded it almost always succeeds.  We set a
job's source to SYNC_CYCLE_CONTINUATION as part of the DOWNLOAD_UPDATES
step, making it very difficult for a job to not have that state by the
time we reach ScheduleNextSync().

The only types of sync jobs that could have reached the else branch are
CLEAR_USER_DATA and CLEANUP_DISABLED_TYPES.  Neither of these types of
jobs are currently at risk of failing (or, if they do fail, we won't
detect or handle it here).

The intention of the else branch seems to have been to handle the first
failure by retrying immediately.  That makes sense, but, since this
branch has been unreachable for so long, I assume that the current
behaviour of starting to delay immediately after the first failure works
well, too.  Therefore, I have decided to modify the code so it better
describes its actual behaviour, rather than modifying the behaviour to
match the intention of the code.

BUG=111229
TEST=


Review URL: http://codereview.chromium.org/9234003