Make sure that the timezone settings are editable by any user.
When the device settings refactoring was done this setting was treated
like any other system setting and made available only to the owner or to
no one if the device was enterprise enrolled.
However this should be made
editable by everyone at least until we get proper system settings delegation
when we might be able to enforce some stricter policy on that setting.

BUG=chromium-os:24450
TEST=As described in the bug.


Review URL: http://codereview.chromium.org/9316093