Add client for background testing of HTTP pipelining.
Note this isn't hooked up, so it won't affect users yet. Once the servers are up, it'll be
hooked up to a field test, much like the UDP echo test. Also missing is a force-pipelining option for requests. I'll add that in a separate CL.

BUG=110794
TEST=unit_tests


Review URL: http://codereview.chromium.org/9302024