[uber extensions] flicker less when enabling/disabling extension
the real fix will be to make the extensions page not clobber several times in a row when an extension is disabled/enabled, but that is more difficult. (See comment in extension_settings_handler.cc::Observe). For now this makes it look a bit better.

BUG=112493
TEST=see bug


Review URL: http://codereview.chromium.org/9381026