[uber] fix lazy load in caching case
the iframe contents can be cached even when src= is not set. The easiest way to tickle this bug is to set chrome to restore session on startup, with chrome://chrome as one of the pages. Workaround: dynamically add the iframe rather than just setting the src attribute (suggested by dbeam)

BUG=none
TEST=none


Review URL: http://codereview.chromium.org/9379031