Fix chrome://omnibox/ to not throw exception when input is empty.
Precisely, the current version of the Javascript throws exceptions
if you check the show-incomplete-results checkbox and also if you play
with any checkboxes after that one is checked -- i.e., forcing a
call to refresh() -- if we don't have any output.  This fixes that.

BUG=
TEST=


Review URL: http://codereview.chromium.org/9385034