Fix about box text spacing on Windows.
This got broken by http://crrev.com/119967 but
was masked out by the use_canvas_skia_skia change.

The issue is that when MULTI_LINE was passed, the
canvas_skia_win.cc path does not count the trailing
space in the string width.

Since this code is doing the wrapping itself, it
should not be passing MULTI_LINE in the first place.

BUG=113744
TEST=Open about dialog on Windows and check that words
have spaces between them.

Review URL: http://codereview.chromium.org/9392007