Fix flakiness in ExtensionFileBrowserPrivateApiTest.FileBrowserMount
First (Unmount) part of the test does some work on the file thread, so we have to make sure we empty the file thread message loop before we end the test.

From test extension's point of view, unmount part of the test ends when GetLocalPathsOnFileThreadAndRunCallbackOnUIThread is posted on the file thread, but for this test to pass, this method's callback has to be called on ui thread.
(The rest of the test is done completely on main thread)

TEST=FileBrowserPrivateApiTest.FileBrowserMount
BUG=114095

Review URL: https://chromiumcodereview.appspot.com/9392018