Refactor FileSystemOperation to take callback for each method.
(Reland of r121620, which was reverted because it broke chromeos-clang compilation and LayoutTests/fast/filesystem/file-writer-*)

This patch is the first step for supporting cross-filesystem copy/move on
the Filesystem API implementation. To accomplish it, I'm planning to
crack FileSystemOperation::{Move,Copy} to a series of other FSO operations.
For it, per-method callback is more handy.

BUG=110121
TEST=*File*
TEST=LayoutTest fast/filesystem

Review URL: http://codereview.chromium.org/9372044