Cocoa: Handle re-open event in app mode
Currently when the Chrome icon is clicked with no windows open we always open a new browser window.

This is incorrect in app mode. In app mode we should instead open an app window.

To fix this I changed the reopen handler to call BrowserInit::LaunchBrowser() which does the correct thing depending on the current mode.

BUG=112651
TEST=


Review URL: http://codereview.chromium.org/9383040