Log V8 Benchmark results to console.
This allows test results to be scraped from the system's console on systems that
support it, such as Android. In that case, it avoids the need to run from the
performance_ui_tests binary.

Also add some missing semicolons at the end of assignment statements
(purely stylistic).

BUG=None
TEST=None


Review URL: http://codereview.chromium.org/9395010