Remove TypedUrlModelAssociator association map.
Removed the association map in TypedUrlModelAssociator because it has been a source
of bugs and the information is redundant.

TypedUrlModelAssociator no longer implements PerDataTypeAssociatorInterface, since
many of the "standard" API implementations in that class are not needed (like
Associate()) or result in inefficient code (such as code that looks up nodes by
their client tag, returns the ID, then looks up the node again by the ID).

BUG=none
TEST=none


Review URL: http://codereview.chromium.org/9350069