Fix bug in HasMetadata: Updates argument even if it fails.
HTTPSRequestTest.HTTPSErrorsNoClobberTSSTest found that we were doing junk
updates. See http://codereview.chromium.org/9233026/.

This CL resolves the flakiness of
HTTPSRequestTest.HTTPSErrorsNoClobberTSSTest: even though there was a real
bug that the test uncovered, the test only caught it occasionally due to
timing factors (checks on the created and expiry fields of the DomainState
object). In fact, since we use a preloaded DomainState, there is no
meaningful time (created or expiry) to check; the time values set in the
DomainState ctor are not meaningful for preloaded DSs.

BUG=110817, 114164
TEST=net_unittests, HTTPSRequestTest.HTTPSErrorsNoClobberTSSTest, TransportSecurityStateTest.*

Review URL: http://codereview.chromium.org/9390030