Adding temp Chrome Endure Gmail test for compose/discard scenario, with a sleep.
The Gmail team requested a test that is the same as the existing compose/discard
Gmail test, except that it sleeps for 10 seconds between each iteration.  This
is to study whether garbage collection "kicks in" by default during that time,
skewing the measured results.  This test is meant to be temporary, to run
side-by-side with the existing compose/discard test, so the Gmail team can
study the effect (if any) of the added sleep.  Later, this test will be
removed or combined with the existing compose/discard Gmail test.

BUG=None
TEST=Verified the test runs on my local linux machine.

Review URL: https://chromiumcodereview.appspot.com/9395029