[Sync] Delay bookmark association until the history db is loaded.
Previously the history db would load in parallel (on the history thread),
resulting in the favicons possibly loading after association has completed.
This is a problem as when they load, they trigger a notification that results
in rewriting all bookmarks, as last seen by the favicon service.

BUG=112518,94941
TEST=unit_tests, sync_integration_tests


Review URL: http://codereview.chromium.org/9348111