Support use of third party time function for about:profiler
This uses TCMalloc to provide a time function, allowing
us to see how much memory was allocated on a single thread
during the running of a task.  The alternate time function
is put in place only when a specific environment variable
is detected during TCMalloc startup.

This change currently is activated only in Windows/Linux,
as it is based on changes TCMalloc (not used on Mac).

We also create an infrastructure for using any alternate
timer, to replace the "wall clock time," on a per-thread
basis, in the about:profiler infrastructure.  That interface
may be used on other platforms, including scenarios where
we have a per-thread-CPU-time function to replace the
wall-clock timer.  In all cases, when this alternate timer
is activated, we lose the ability to calculated queueing
time.  Queueing time is based on a time snapshot taken on
a second thread, and hence is not comparable to the
alternate timer (when the alternate timer is engaged).


r=rtenneti
BUG=103321
Review URL: https://chromiumcodereview.appspot.com/9212025