Merge 121310 - Make the Chrome Web Store Icon Syncable
Mark the CWS as a syncable app so that it page and app launch ordinals are synced. It also ensure that the CWS has valid ordinals as soon as it it added to the client, instead of waiting until the NTP is shown.

BUG=112290
TEST=Users should be able to move the Chrome Web store icon and have its positions update to other signed in locations.


Review URL: http://codereview.chromium.org/9340007

TBR=csharp@chromium.org
Review URL: https://chromiumcodereview.appspot.com/9417018