Windows Breakpad: Get metadata from chrome.exe instead of chrome.dll
Comments in the code indicate that getting the metadata from
chrome.dll in GetCustomInfo() is slow because the DLL might not have
loaded at the point where Breakpad is initialised.  Getting the
metadata from the current module (which is either chrome.exe or
nacl64.exe) will be faster since the module is already loaded.

This will let us initialise Breakpad on the main thread in a later
change.

I tested this by adding the following debugging prints:

  printf("product_version=%ls\n", version_info->product_version().c_str());
  printf("product_short_name=%ls\n",
version_info->product_short_name().c_str());
  printf("is_official_build=%i\n", version_info->is_official_build());
  printf("special_build=%ls\n", version_info->special_build().c_str());

This produced the same output before and after the change:

  product_version=19.0.1041.0
  product_short_name=Chromium
  is_official_build=0
  special_build=

and it produced the same from nacl64.exe when NaCl was run.  These are
all the fields of version_info that this code uses.

We rename "dll_path" to "exe_path" for clarity.

The other use of "dll_path", for IsPerUserInstall(), is safe to change
to use chrome.exe's path because this only looks at the directory of
the pathname.

We change chrome.exe to use the same method (GetModuleFileNameW()) as
nacl64.exe for getting the current executable's pathname.

BUG=111265
TEST=see above

Review URL: https://chromiumcodereview.appspot.com/9404034