Add UMA stats for save package duplicate file errors.
Sometimes the save package code will try to save multiple frames
with the same file name. In this case the first frame will win.
Its not clear how often this happens; this change adds UMA stats
so we can get a feel for this.

BUG=106364
TEST=Checked save package still works, and the UMA code is
     triggered on the known problem site.


Review URL: http://codereview.chromium.org/8805002