Merge 121878 - Allow SessionBackend::Init() to occur just-in-time in production builds.
This fixes a regression introduced in r113377 that caused the "old
session" to be switched at start-up before being used to populate the
recently closed menu.

This is the simplest possible fix, intended for merge to M18.  A more
robust fix that prevents future regressions is likely possible and may
be investigated as a follow-up.

BUG=110785
TEST=see bug repro steps

Review URL: https://chromiumcodereview.appspot.com/9361056

TBR=joi@chromium.org
Review URL: https://chromiumcodereview.appspot.com/9419049