Add Omnibox field trial for inlining HistoryQuickProvider suggestions.
Creates a field trial to control whether it's okay to inline 
HistoryQuickProvider suggestions.

Adds code to the history provider classes to implement this change.

Adds a third-value command line flag for this change: auto, allowed, 
prohibited.

Adds an about:flag control for this change.

Adds an UMA histogram to log the state of the field trial.

BUG=112844
TEST=By hand


Review URL: http://codereview.chromium.org/9370027