Adds dump accessibility tree support in Windows.
This largely adds human readable strings for IAccessible2 roles and states.
This is in part to make it possible to hand-edit the expected files we generate.
Sidenote: switch to string16 usage from std::string as much as possible (despite the incompatibility with API's provided by FilePath).

Review URL: https://chromiumcodereview.appspot.com/9295031