Update the format of the Pepper Flash manifest.
Allow multiple "or" possibilities for interfaces. E.g., a string of the form
"Foo|Bar|Baz" indicates that (at least) one of the interfaces "Foo", "Bar",
"Baz" must be supported (for the plugin to be compatible).

This is needed so that we can remove support for old interfaces.


Review URL: http://codereview.chromium.org/9424018