net: Simplify FileStream implementation on POSIX.
This is a re-land of crrev.com/122780, which caused memory corruption,
and was reverted. Valgrind bots caught invalid memory access.
The problem was as simple as not holding the IOBuffer, until read/write
was done.

Replace AsyncContext with PostTaskAndReply and WeakPtrFactory.
This patch shaves about 150 lines.

This patch also fixes a subtle threading issue with the POSIX
implementation. Callbacks passed to Read/Write() should be
run on the thread where Read/Write() were called.

However, before this patch, callbacks passed to Read/Write were
actually run on the thread where the async context object was created.

BUG=114783
TEST=try bots including valgrind bots.

Review URL: https://chromiumcodereview.appspot.com/9355054