Make offline error page visible
The problem happened because the page used chrome:// to load local_strings.js but it failed because page itself is served from data URL that has no access to chrome://. Error in the log was "Not allowed to load local resource". Now local_strings.js is flatten and it fixes the issue. Also return timeout for showing offline error page.

BUG=chromium-os:23792
TEST=manual


Review URL: http://codereview.chromium.org/9427008