Avoid profiler initialization in exe context
The profiler is "really" only useful in the
DLL, and there was one lingering Thread::SetName
call that (accidentally) initialized the
profiler in the exe context.  This change
avoids starting the profiler for that thread
which is only part of the exe, and is never used
for spawning or running tasks (profiled items).

r=rvargas
bug=103321
Review URL: https://chromiumcodereview.appspot.com/9432011