chromeos: bluetooth: complete the BlueZ API
Implement the remaining set of properties, methods and signals on the
Manager, Adapter and Device client objects to fully implement the
BlueZ API for when we need it.
    
Added callbacks to the existing StartDiscovery and StopDiscovery
Adapter methods to make them consistent. While they can only return
to indicate success or error, failing to enter discovery mode will
be an important thing to know about.
    
Also for consistency dropped log messages to WARNING in the case of
receiving the wrong response types, and in the Manager Stub actually
call callbacks to indicate failure rather than leave the client
hanging.

BUG=chromium-os:22087
TEST=tested on client device

Change-Id: Ifc38c1f29b22c42880d4a81a825f128907cac4ec


Review URL: http://codereview.chromium.org/9372089