Globbing _manifest_features.json out of the extension documentation.
Adding this file broke the extension documentation build script, as it reads
_all_ the JSON files in the directory, and attempts to treat them all as API
docs. The fix I'm introducing in this CL is a stopgap, just so folks can
continue to build docs in the meantime. I don't expect it's the right solution,
just good enough for the moment.

BUG=115614
TEST=run build.py: no errors.
TBR=kalman@chromium.org

Review URL: https://chromiumcodereview.appspot.com/9464005