Set minimize property and re-enable tests on Aura
Currently we disable several important tests on Aura because Minimize() is unimplemented.
Most significant is ExtensionApiTest.UpdateWindowResize and ExtensionApiTest.UpdateWindowState, which test a lot of the chrome.windows api.

Setting the minimized property but not doing anything satisfies the tests with minimal change, until we determine what the expected minimize behavior is on Aura, and how this affects the extensions API.

BUG=104571
TEST=Tests pass (esp. browser_tests)


Review URL: http://codereview.chromium.org/9372126