[Mac] Implement logging:DumpWithoutCrashing().
Bring OSX up to parity with other platforms.  It is currently in use
in the thread-watcher code.

Use BreakpadGenerateAndSendReport() to generate a crash report without
actually crashing.  This function does not dump the current thread's
stack, so spin up a separate thread to do the dump, then wait on that
thread to complete.

BUG=none
TEST=Monitor for thread-watcher dumps on the crash server.


Review URL: http://codereview.chromium.org/9460028