Fix PasswordStoreWinTest.Migration leak.
This leak was introduced in http://crbug.com/83999 by forgetting to call
store->Shutdown() at the end of the Migration test.  I made the
scoped_refptr<PasswordStore> a member of the fixture and put the
store_->Shutdown() calls in the TearDown method.

This leak was holding open handles to db files, which was also
preventing the cleanup of the db's temp directory on Windows.

R=sky@chromium.org,timurrrr@chromium.org
BUG=115191
TEST=PasswordStoreWinTest.* under drmemory gives no leaks, no temp dir warning


Review URL: http://codereview.chromium.org/9464022