Fix new tab button border
In Aero mode the border of the new tab button was hard to see.

The problem was that I was drawing the border then masking out the bitmap. This caused parts of the border to be removed.

To fix this I'm masking out the image then drawing the border.

Screenshots:
  old button: http://imgur.com/a/juCLJ
  new button: http://imgur.com/a/Fc1jH

BUG=
TEST=


Review URL: http://codereview.chromium.org/9475020