Fix bug where we don't run all the browser_tests when using gtest_filter={}
We are reaching a max limit in the shell string size of 32k characters when
specifying the 1200~ or so browser_tests using gtest_filter. In fact because
we are specifying approximately 57k of characters for tests, we're roughly
only running 1/2 of the browser_tests.

In order to fix this we move to running a sub-group of the tests at a time.
gtest_runner which parses chrome test output back into individual tests doesn't
care about gtest being reinvoked so this will appear the same on the test
dashboard.

In this CL I also fix the chrome test suite to actually use the number of
machines it is given. This allows us to pass count!=1 for browser_tests and
actually have it make a difference.

Original change reviewed in:
https://gerrit.chromium.org/gerrit/#change,16339

BUG=chromium-os:26487, chromium-os:26697
TEST=workon autotest, autotest-chrome, ran control.desktopui_browsertest and
inspected output.

Change-Id: I2e6a68ee6389799694fe685c9d2db3e97d4c77b8

Review URL: https://chromiumcodereview.appspot.com/9455025