gdata: Cut inheritance from DocumentsService to GDataService.
The inheritance got in the way for achuith and satorux. :)
Composition is cleaner here, as GDataService is essentially
an authentication service, whereas DocumentsService is just
a client of the authentication service.

GDataService is renamed to GDataAuthService.
SupportsWeakPtr<> is replaced with WeakPtrFactory along the way,
in favor of less inheritance.

BUG=none
TEST=the file manager worked as before (i.e. gdata files are listed)

Review URL: https://chromiumcodereview.appspot.com/9559006