Working around http://code.google.com/p/gyp/issues/detail?id=13
Visual Studio doesn't guarantee the 'copies' stanza is executed after the grit target, which can lead to a failure if it tries to copy the file before it's been generated. This puts the grit action in its own target, and making the copies stanza its own target that depends on the grit target.

Review URL: http://codereview.chromium.org/9558003