[Sync] Add command line switch to disable syncing open tabs.
Because open tabs is not enabled by default we rely on the kEnableSyncTabs flag
to turn it on. But, because that flag can be set by way of sync logic, we need
some way to have a user disable syncing open tabs via the command line (in case
of tab sync causing crashes or something).

BUG=none
TEST=none


Review URL: http://codereview.chromium.org/9549027