Begin to separate the MetricsService logic for creating vs uploading logs
The biggest change here is to remove the dedicated current->staged->upload pipeline, and instead have finishing a log add it to the queue (the same one that's used for old logs that came from disk), and always do uploading by staging from the queue. That will allow for MetricsService to handle cutting logs and uploading logs separately in a future CL (and allows for a bit of simplification now in the shutdown case, where it's no longer necessary to stage the current log for upload just so that it's possible to immediately store it).

The rest of the MetricsService changes aren't actually changing the flow, the just split some things into more methods so that future CLs to separate cutting logs and uploading them will be smaller and easier to understand.

StartFinalLogInfoCollection/OnFinalLogInfoCollectionDone were added to make it much more clear how StartScheduledUpload gets to the code that actually does the uploading, since it's confusing to have to follow the jumps through the various implementation details of the specific info being gathered to see what the end point is.

BUG=None
TEST=Metrics should continue to be reported from official builds


Review URL: http://codereview.chromium.org/9396001