gdata: Fix a crash at deletion of GDataFileSystem on debug build.
GDataFileSystem was a ref counted object, but deleted by
ProfileKeyedServiceFactory with 'delete'. This was horribly
wrong, and resulted in a crash at deletion of GDataFileSystem
that happened at browser shutdown time.

The fix is to make GDataFileSystem non-ref counted.
WeakPtr is used to create callbacks.

Along the way, replaced kGDataAPICallThread with BrowserThread::UI
in gdata.cc. TokenFetcher requires callers to be on UI thread.

BUG=chromium-os:27121
TEST=Run chrome on debug build, open the file manager, show gdata files, Ctrl-C to exit. No crash from GDataFileSystem

Review URL: https://chromiumcodereview.appspot.com/9564002