Prefix Android-string descriptions with Mobile:
Some strings currently have a specific version for Android, but the
description attribute does not mention this.

This prefixes "Mobile: " for all Android strings where the same message
name is already in use, and Android just has a specific version of a
string for mobile.

BUG=114046
TEST=Verify that strings show up correctly in Android build.


Review URL: http://codereview.chromium.org/9567017