gdata:: Add GDataFileSystem::GetFile().
This function is used to get files from the file system.
For now, it always download files from the cloud.
We should have the caching layer in-pace here.

Along the way, a mssing 'return' to Remove().
Change GetDocumentUrlFromPath() to GetGDataFileInfoFromPath,
and wrote a test for it.

BUG=chromium-os:27174
TEST=unit_tests


Review URL: http://codereview.chromium.org/9583031