ash uber tray: Allow customizing each item depending on whether the user is logged in or not.
If a user is not logged in (e.g. in the login screen, or during an incognito
session), some of the items need to behave differently. So provide this
information to each of the items so that such customization is possible. Also,
make sure all the items get a chance to update themselves when user logs in.

BUG=110130
TEST=tested on device, logged in as normal user, and incognito.

Review URL: https://chromiumcodereview.appspot.com/9580024