Implement registry of ongoing GData operations.
This patch adds an way to manage all the GData operations on-flight. Specifically,
- CancellAll() for canceling all the operations.
- SetObserver() for observing the start/progress/finish of the active operations.

Currently not enough information (about which operation corresponds to which) is passed to the observer for end-user UI.
I rather want to adjust it later after UI layer is matured and what we need is determined.

BUG=chromium-os:26907
TEST=unit_tests --gtest_filter='*GData*'
(test for actual gdata operations are to come...)

Review URL: https://chromiumcodereview.appspot.com/9564003
Patch from Kazuhiro Inaba <kinaba@chromium.org>.